adjust generateProblems, move files
